Fix typo: rename registation-form to registration-form
Corrects misspelling in component directory name and updates import.
