+++ frontend/courses-front/src/components/registration-form/index.jsx
