+++ frontend/courses-front/src/components/registration-form/styled.js
